finnish
将todo-list和用户结合

+++ routes_todo.py
@@ -1,6 +1,7 @@
 from utils import log
 from models.user import User
 from models.todo import Todo
+from routes import session
 
 
 def response_with_headers(headers, code=200):
@@ -10,7 +11,7 @@ def response_with_headers(headers, code=200):
     """
     header = 'HTTP/1.1 {} OK\r\n'.format(code)
     header += ''.join(['{}: {}\r\n'.format(k, v)
-                           for k, v in headers.items()])
+                       for k, v in headers.items()])
     return header
 
 
@@ -19,6 +20,7 @@ def redirect(path):
         'Location': path,
     }
     return response_with_headers(headers, 302).encode('utf-8')
+
 
 def template(path):
     path = 'templates/' + path
@@ -30,19 +32,27 @@ def route_todo(request):
     """
     todo_list 首页的路由函数
     """
+    log("cookies", request.cookies)
+    session_id = request.cookies.get('user')
+    log('route todo session_id and session', session_id, session)
+    u_id = session.get(session_id, -1)
+    if u_id == -1:
+        return redirect('login')
     headers = {
         'Content-Type': 'text/html',
     }
-    #得到所有的todo
+    # 得到所有的todo
     todo_list = Todo.all()
     todos = []
     i = 1
+    # 注入属于用户的todo
     for t in todo_list:
-        delete_href = '<a href="/todo/delete?id={}">删除</a>'.format(t.id)
-        edit_href = '<a href="/todo/edit?id={}">编辑</a>'.format(t.id)
-        s = '<h3>{} : {}  {}  {}</h3>'.format(i, t.title, delete_href, edit_href)
-        i = i + 1
-        todos.append(s)
+        if t.id == u_id:
+            delete_href = '<a href="/todo/delete?id={}">删除</a>'.format(t.id)
+            edit_href = '<a href="/todo/edit?id={}">编辑</a>'.format(t.id)
+            s = '<h3>{} : {}  {}  {}</h3>'.format(i, t.title, delete_href, edit_href)
+            i = i + 1
+            todos.append(s)
     todo_html = ''.join(todos)
     # 替换模板文件中的标记字符串
     body = template('todo.html')
@@ -74,9 +84,15 @@ def route_add_todo(request):
     """
     增加一个todo的处理函数
     """
-    #得到所有的todo
+    # 得到用户信息
+    session_id = request.cookies.get('user')
+    u_id = session.get(session_id, -1)
+    if u_id == -1:
+        return redirect('login')
     form = request.form()
     todo = Todo.new(form)
+    # 增加user_id
+    todo.user_id = u_id
     todo.save()
     return redirect('/todo')
 
@@ -85,10 +101,16 @@ def route_update_todo(request):
     """
     删除一个todo的处理函数
     """
-    #得到todo对象
+    # 得到用户信息
+    session_id = request.cookies.get('user')
+    log('route todo session_id and session', session_id, session)
+    u_id = session.get(session_id, -1)
+    # 得到todo对象
     form = request.form()
     todo_id = int(form.get('id', None))
     t = Todo.find_by(id=todo_id)
+    if t.user_id != u_id:
+        return redirect('/todo')
     t.title = form.get('title', '')
     t.save()
     return redirect('/todo')
@@ -98,13 +120,14 @@ def route_delete_todo(request):
     """
     删除一个todo的处理函数
     """
-    #得到todo对象
+    # 得到todo对象
     todo_id = int(request.query.get('id', None))
     t = Todo.find_by(id=todo_id)
-    #删除
+    # 删除
     if t is not None:
         t.remove()
     return redirect('/todo')
+
 
 route_todo_dict = {
     '/todo': route_todo,
